--- src/main/java/CarGame2/InvalidInputException.java
@@ -1,8 +1,0 @@
-package CarGame2;
-
-public class InvalidInputException extends RuntimeException {
-
-    public InvalidInputException(String message) {
-        super(message);
-    }
-}
