+++ src/main/java/CarGame2/domain/model/Cars.java
@@ -1,29 +1,29 @@
-package CarGame2;
+package CarGame2.domain.model;
 
+import CarGame2.domain.service.Generator;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Cars {
 
     private final List<Car> cars;
-    private final Generator generator;
 
-    public Cars(List<String> carNames, Generator generator) {
-        this.generator = generator;
+    public Cars(List<String> carNames) {
         this.cars = createCars(carNames);
     }
 
     private List<Car> createCars(List<String> carNames) {
         List<Car> carsList = new ArrayList<>();
         for (String name : carNames) {
-            carsList.add(new Car(name, generator));
+            carsList.add(new Car(name));
         }
         return carsList;
     }
 
-    public void moveAll() {
+    public void moveAll(Generator generator) {
         for (Car car : cars) {
-            car.move();
+            int speed = generator.generate();
+            car.move(speed);
         }
     }
 
@@ -46,17 +46,6 @@ public class Cars {
         return car.getPosition() == maxPosition;
     }
 
-    public void printCarStatuses(OutputHandler outputHandler) {
-        for (Car car : cars) {
-            outputHandler.printCarStatus(car);
-        }
-        outputHandler.printRoundEnd();
-    }
-
-    public List<Car> getCars() {
-        return cars;
-    }
-
     public int getMaxPosition() {
         int maxPosition = 0;
         for (Car car : cars) {
@@ -64,4 +53,9 @@ public class Cars {
         }
         return maxPosition;
     }
+
+    public List<Car> getCars() {
+        return cars;
+    }
 }
+
