+++ src/main/java/CarGame2/domain/service/RandomGenerator.java
@@ -1,13 +1,14 @@
-package CarGame2;
+package CarGame2.domain.service;
 
 import java.util.Random;
 
 public class RandomGenerator implements Generator {
 
+    private static final int MAX_RANDOM_NUMBER = 10;
     private final Random random = new Random();
 
     @Override
     public int generate() {
-        return random.nextInt(10);
+        return random.nextInt(MAX_RANDOM_NUMBER);
     }
 }
