+++ src/test/java/RaceTest.java
@@ -1,9 +1,8 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
-import CarGame2.Car;
-import CarGame2.Generator;
-import CarGame2.OutputHandler;
-import CarGame2.Race;
+import CarGame2.domain.model.Car;
+import CarGame2.domain.model.Race;
+import CarGame2.domain.service.Generator;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -22,7 +21,7 @@ class RaceTest {
                 return 5;
             }
         };
-        race = new Race(List.of("car1", "car2", "car3"), 3, fixedGenerator, new OutputHandler());
+        race = new Race(List.of("car1", "car2", "car3"), 3, fixedGenerator);
     }
 
     @Test
